Make vkb work on Tizen emulator
VKB does not pop-up on Tizen emulator because mouse events are not
handled. A focused window handle should be passed to
WaylandInputDevice in WaylandPointer::OnPointerEnter, but it is not
called on Tizen emulator. Instead, touch events are fired. So this
patch allows to set window focus when the window shows or touch
events are fired.

Bug=#334